refactor: merge memory_tier headers
