+++ include/blender/base_types.h
@@ -6,7 +6,7 @@
 // Include result definitions and math compatibility
 #include "core/common.h"
 #include "cuda/math_common.h"
-#include "memory/memory_tier.h"
+#include "memory/memory_tier.hpp"
 
 #include <glm/glm.hpp>
 #include <cstdint>
