+++ include/memory/unified_memory.h
@@ -4,7 +4,7 @@
 #include "cuda/cuda_defs.h"
 #include "memory/manager.h"
 #include "memory/types.h"
-#include "memory/memory_tier.h"
+#include "memory/memory_tier.hpp"
 #include <cstddef>
 
 namespace sep {
