+++ include/quantum/quantum_pattern_processor.h
@@ -2,7 +2,7 @@
 
 #include "core/system_hooks.h"
 #include "core/types.h"
-#include "memory/memory_tier.h"
+#include "memory/memory_tier.hpp"
 #include "pattern/pattern_evolution.h"
 #include "quantum/quantum_processor.h"
 #include "quantum/types.h"
