+++ src/memory/memory_tier_manager.cpp
@@ -3,7 +3,6 @@
 #include "core/types.h"
 #include "cuda/kernels.cuh"
 #include "memory/logger.hpp"
-#include "memory/memory_tier.h"
 #include "memory/memory_tier.hpp"
 #include "memory/memory_tier_manager.hpp"
